Use multiprocessing.pool instead of JoinableQueue
With JoinableQueue I had to manage workers manually. If one died I had
to catch that signal and start a new one.

With Pool, I have to deal with workers adding more work to the "queue"
after I've closed and joined the pool. Specifically, syncing the org
from the command line generates a job that generates other jobs.

Pool manages the workers for me.